+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- allow to pretty print result
+
 ## [1.0.4] - 2024-04-09
 
 ### Removed
@@ -16,5 +22,6 @@
 - Fix: Reading from empty file does not throw anymore
 - Prepare publish
 
+[Unreleased]: https://github.com/inlavigo/gg_json/compare/1.0.4...HEAD
 [1.0.4]: https://github.com/inlavigo/gg_json/compare/1.0.3...1.0.4
 [1.0.3]: https://github.com/inlavigo/gg_json/tag/%tag